Tried manually splitting up a module; bad idea do not try again [ci skip]
It's manual work which the computer can do for us, as well as doing a
better job (in its access to the docstrings and things it can create).

This was an attempt to see if we can avoid using astropy-helpers; I
think I'll try that next (as an external requirement that isn't
specified as I don't want to waste time getting the astropy boot
code integrated if we don't use it).