Revert to numeric numbering of references
It's not entirely clear to me whether you can use non-numeric labels
for references; it does seem to work but I have seen problems with the
documentation tools being unable to find the reference. Now, this may
be because of other issues (e.g. it's a docstring from compiled code
perhaps), but switch to all-numeric references for now.